requirements.txt
pandas
joblib
torch===1.4.0  --find-links https://download.pytorch.org/whl/torch_stable.html
torchvision===0.5.0  --find-links https://download.pytorch.org/whl/torch_stable.html
numpy==1.18.1
scipy==1.3.2
tqdm==4.41.1
pytorch==1.4.0
seaborn==0.10.0
ipykernel
scikit-learn==0.22.1
spotipy
python-dotenv